Update natives and documentation
Instead of cloning the handle we copy the whole KeyValues struct into a new instance.

This means that the original copy is immutable (the plugin can't pull the KV and modify it as is), which with documentation provides more freedom in how the API internally stores attributes (we might switch to StringMap or similar later).